Various changes... including adding a directionButtons mixin for those slider controls...  I'll deal with the Javascript for it later...
Fixed the formatting of the socialMediaIcons div however it takes 2 reloads for people to see that it's working... weird...

Used the .one-third class by adding it to the footer and indenting the include under it.  Since div was set to margin and padding 0 at the very start there shouldn't be any worry about ghost space.  Used a mixing for popularPosts to avoid repeating code over long stretches.

Added a "flex-direction-nav" to sliderImage however it's set to display:none until I figure out how to work the Javascript later in the final phases.