I started doing the responsive coding for this and realized that I may need Javascript to make this work as the CSS trick to create a pop out nav menu gets complicated because sidebarNav is already position relative and I need the responsive nav menu icon to be positioned relative to the body.
Well I suppose I could position it relative to the sidebar however that means its positioning must be in percentages otherwise as the screen shrinks I'll have to constantly re-update its position.

At this point I'm still inserting the code to either construct an alternate nav menu or use the existing one.

If I'm going to use the existing one then I need to increase its width or change it from position fixed to position absolute because when I looked at it on my real iPhone screen (or the iPhone simulator in Chrome) the lower half of the sidebar is cut off and can't be reached because it is fixed.